sort - filter - page

+++ controllers/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/categoryModel')
+const Products = require('../models/productModel')
 
 const categoryController = {
     getCategories: async (req,res) => {
@@ -31,6 +32,8 @@ const categoryController = {
     },
     deleteCategory: async (req, res) => {
         try {
+            const products = await Products.findOne({category: req.params.id})
+            if(products) return res.status(400).json({status: false, msg:"Warning: Category have products. Please delete products first !"})
             await Category.findByIdAndDelete(req.params.id);
             res.json("Delete success");
         } catch (error) {
